refactor: extract ui onboarding helpers
